fix(current_flow): return Python float from current_flow_closeness/information_centrality (br-r37-c1-ltkoo)
current_flow_closeness_centrality (and its alias information_centrality) built
the result dict from a numpy array (1.0 / centrality_arr[i]), yielding
np.float64 values where networkx returns Python float. np.float64 subclasses
float so isinstance passes, but type(v) is float fails and repr differs
(np.float64(...)). current_flow_betweenness_centrality already coerces to
float; this aligns the closeness variant. Values unchanged; 50 existing
current-flow tests pass. Found via a return-type exactness probe.

+++ python/franken_networkx/__init__.py
@@ -25387,7 +25387,11 @@ def current_flow_closeness_centrality(
         centrality_arr[node] += n * diag - 2.0 * float(row.sum())
         centrality_arr += diag
 
-    return {ordering[i]: 1.0 / centrality_arr[i] for i in range(n)}
+    # br-r37-c1-cfctype: nx returns Python ``float`` dict values; ``centrality_arr``
+    # is a numpy array so ``1.0 / centrality_arr[i]`` is ``np.float64``, which
+    # diverges on ``type(v) is float`` and on repr (``np.float64(...)``).
+    # current_flow_betweenness already coerces — match it here.
+    return {ordering[i]: float(1.0 / centrality_arr[i]) for i in range(n)}
 
 
 def betweenness_centrality_subset(
